+++ src/hooks/index.js
@@ -18,6 +18,7 @@ export { default as useUpdateOwnership } from './useUpdateOwnership';
 export { default as useLocalStorageItems } from './useLocalStorageItems';
 export { default as useItemAuditDataQuery } from './useItemAuditDataQuery';
 export { default as useInventoryVersionHistory } from './useInventoryVersionHistory';
+export { default as useTotalVersions } from './useTotalVersions';
 export * from './useQuickExport';
 export * from '@folio/stripes-inventory-components/lib/queries/useInstanceDateTypes';
 export * from './useCallNumberTypesQuery';