+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-inventory
 
-## [13.0.0] (IN PROGRESS)
+## [13.0.0](https://github.com/folio-org/ui-inventory/tree/v13.0.0) (2025-03-14)
+[Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.12...v13.0.0)
 
 * ECS | Instance details pane does not contain all tenants when user does not have affiliations / permissions in all tenants. Fixes UIIN-3113.
 * *BREAKING* Rename quick-marc routes from `bib` to `bibliographic` and `duplicate` to `derive`. Refs UIIN-3120.
@@ -52,6 +53,10 @@
 * Make user name hyperlink in version history inactive if user does not have permissions. Refs UIIN-3269.
 * Display Original version card of the audit log with no field changes. Refs UIIN-3270.
 * Change heading of modal generate accession and call number. Refs UIIN-3274.
+* ViewSource - add tenantId to useAuditSettings. Refs UIIN-3266.
+* Enhancement help text on Settings > Inventory > Number generator options. Refs UIIN-3217.
+* Add default value for number of versions counter on "Version history" pane. Fixes UIIN-3283.
+* Use `stripes.okapi.tenant` in order to get persistent id between pages. Fixes UIIN-3251.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.11...v12.0.12)